[Refactor] Address WorldModelEnv review feedback
Per @vmoens review comments on the WorldModelEnv PR:

- WorldModelEnv:
  - Annotate `world_model` parameter as `WorldModel` (via TYPE_CHECKING
    to avoid the torchrl.modules <-> torchrl.envs circular import).
  - Replace `.clone(recurse=False)` with `.copy()`.
  - Drop unnecessary `{}` in `TensorDict(batch_size=..., device=...)`.
  - Iterate spec leaves with `include_nested=True, leaves_only=True`
    so nested observation/reward/done specs survive `_step` end-to-end.
  - Use `full_*_spec.zero()` (and `full_*_spec[key].zero()`) instead of
    hand-rolled `torch.zeros(...)` for default reward/done/terminated
    tensors, keeping shapes / dtypes consistent with the specs.
- WorldModelLoss: drop unused `batch_size=[]` from final TensorDict
  construction.
- Tests: consolidate `test/test_world_model.py` (deleted) into:
  - `test/envs/test_model_based.py` — TestWorldModelForward,
    TestWorldModelEnv (and a `_SpecOnlyEnv` helper).
  - `test/objectives/test_dreamer.py` — TestWorldModelLoss.

No public-API changes. 22/22 affected tests pass locally.
