+++ test/envs/test_model_based.py
@@ -9,10 +9,16 @@ import re
 import numpy as np
 import pytest
 import torch
-from torch import nn
 
+from tensordict import TensorDict
+from tensordict.nn import TensorDictModule
+from torch import nn
+from torchrl.data import Composite, TensorDictReplayBuffer, Unbounded
+from torchrl.data.tensor_specs import Categorical
+from torchrl.envs.common import EnvBase
+from torchrl.envs.model_based import WorldModelEnv
 from torchrl.envs.utils import check_env_specs
-from torchrl.modules import SafeModule
+from torchrl.modules import SafeModule, WorldModel
 from torchrl.modules.tensordict_module import WorldModelWrapper
 from torchrl.testing import get_default_devices
 from torchrl.testing.mocking_classes import ActionObsMergeLinear, DummyModelBasedEnvBase
@@ -105,3 +111,265 @@ class TestModelBasedEnvBase:
         mb_env.step(td)
         # we should be able to do a step with a tensordict that has been expended
         mb_env.step(td_expanded)
+
+
+# ---------------------------------------------------------------------------
+# WorldModel + WorldModelEnv tests
+# (moved here from test/test_world_model.py per review consolidation request)
+# ---------------------------------------------------------------------------
+
+_WM_OBS_DIM = 8
+_WM_LATENT_DIM = 4
+_WM_ACTION_DIM = 2
+_WM_BATCH = 3
+
+
+class _CatLinear(torch.nn.Module):
+    """Concatenates all positional inputs along the last dim, then applies Linear."""
+
+    def __init__(self, in_features: int, out_features: int) -> None:
+        super().__init__()
+        self.linear = torch.nn.Linear(in_features, out_features)
+
+    def forward(self, *tensors: torch.Tensor) -> torch.Tensor:
+        return self.linear(torch.cat(tensors, dim=-1))
+
+
+def _make_linear_world_model(
+    with_done_head: bool = False, with_decoder: bool = False
+) -> WorldModel:
+    encoder = TensorDictModule(
+        torch.nn.Linear(_WM_OBS_DIM, _WM_LATENT_DIM),
+        in_keys=["observation"],
+        out_keys=["latent"],
+    )
+    dynamics = TensorDictModule(
+        _CatLinear(_WM_LATENT_DIM + _WM_ACTION_DIM, _WM_LATENT_DIM),
+        in_keys=["latent", "action"],
+        out_keys=[("next", "latent")],
+    )
+    reward_head = TensorDictModule(
+        torch.nn.Linear(_WM_LATENT_DIM, 1),
+        in_keys=[("next", "latent")],
+        out_keys=[("next", "reward")],
+    )
+    done_head = None
+    if with_done_head:
+        done_head = TensorDictModule(
+            torch.nn.Linear(_WM_LATENT_DIM, 1),
+            in_keys=[("next", "latent")],
+            out_keys=[("next", "done")],
+        )
+    decoder = None
+    if with_decoder:
+        decoder = TensorDictModule(
+            torch.nn.Linear(_WM_LATENT_DIM, _WM_OBS_DIM),
+            in_keys=["latent"],
+            out_keys=["reconstructed_observation"],
+        )
+    return WorldModel(
+        encoder, dynamics, reward_head, done_head=done_head, decoder=decoder
+    )
+
+
+def _make_start_td(batch: int = _WM_BATCH) -> TensorDict:
+    wm = _make_linear_world_model()
+    td = TensorDict(
+        {"observation": torch.randn(batch, _WM_OBS_DIM)}, batch_size=[batch]
+    )
+    return wm.encode(td)
+
+
+def _constant_policy(
+    latent_dim: int = _WM_LATENT_DIM, action_dim: int = _WM_ACTION_DIM
+):
+    return TensorDictModule(
+        torch.nn.Linear(latent_dim, action_dim),
+        in_keys=["latent"],
+        out_keys=["action"],
+    )
+
+
+class _SpecOnlyEnv(EnvBase):
+    """Minimal :class:`EnvBase` used to supply specs to :class:`WorldModelEnv`.
+
+    Not steppable on its own — only its specs are read by the env wrapper.
+    """
+
+    def __init__(self, batch_size: int = _WM_BATCH, device: str = "cpu") -> None:
+        super().__init__(batch_size=[batch_size], device=device)
+        self.observation_spec = Composite(
+            latent=Unbounded(shape=(batch_size, _WM_LATENT_DIM), device=device),
+            shape=[batch_size],
+            device=device,
+        )
+        self.action_spec = Unbounded(shape=(batch_size, _WM_ACTION_DIM), device=device)
+        self.reward_spec = Unbounded(shape=(batch_size, 1), device=device)
+        self.done_spec = Categorical(
+            n=2, shape=(batch_size, 1), dtype=torch.bool, device=device
+        )
+
+    def _reset(self, tensordict, **kwargs):  # pragma: no cover - never called
+        raise NotImplementedError
+
+    def _step(self, tensordict):  # pragma: no cover - never called
+        raise NotImplementedError
+
+    def _set_seed(self, seed):  # pragma: no cover - never called
+        return seed
+
+
+class TestWorldModelForward:
+    def test_output_keys_present(self):
+        wm = _make_linear_world_model()
+        td = TensorDict(
+            {
+                "observation": torch.randn(_WM_BATCH, _WM_OBS_DIM),
+                "action": torch.randn(_WM_BATCH, _WM_ACTION_DIM),
+            },
+            batch_size=[_WM_BATCH],
+        )
+        out = wm(td)
+        assert ("next", "latent") in out.keys(include_nested=True)
+        assert ("next", "reward") in out.keys(include_nested=True)
+
+    def test_encode_shortcut(self):
+        wm = _make_linear_world_model()
+        td = TensorDict(
+            {"observation": torch.randn(_WM_BATCH, _WM_OBS_DIM)},
+            batch_size=[_WM_BATCH],
+        )
+        encoded = wm.encode(td)
+        assert "latent" in encoded.keys()
+
+    def test_step_shortcut(self):
+        wm = _make_linear_world_model()
+        td = _make_start_td()
+        td["action"] = torch.randn(_WM_BATCH, _WM_ACTION_DIM)
+        stepped = wm.step(td)
+        assert ("next", "latent") in stepped.keys(include_nested=True)
+
+    def test_decode_shortcut(self):
+        wm = _make_linear_world_model(with_decoder=True)
+        td = _make_start_td()
+        decoded = wm.decode(td)
+        assert "reconstructed_observation" in decoded.keys()
+
+    def test_decode_without_decoder_raises(self):
+        wm = _make_linear_world_model(with_decoder=False)
+        td = _make_start_td()
+        with pytest.raises(RuntimeError, match="decoder"):
+            wm.decode(td)
+
+    def test_nested_latent_key(self):
+        """A WorldModel composed with nested latent keys still works end-to-end."""
+        encoder = TensorDictModule(
+            torch.nn.Linear(_WM_OBS_DIM, _WM_LATENT_DIM),
+            in_keys=["observation"],
+            out_keys=[("encoded", "latent")],
+        )
+        dynamics = TensorDictModule(
+            _CatLinear(_WM_LATENT_DIM + _WM_ACTION_DIM, _WM_LATENT_DIM),
+            in_keys=[("encoded", "latent"), "action"],
+            out_keys=[("next", "encoded", "latent")],
+        )
+        reward_head = TensorDictModule(
+            torch.nn.Linear(_WM_LATENT_DIM, 1),
+            in_keys=[("next", "encoded", "latent")],
+            out_keys=[("next", "reward")],
+        )
+        wm = WorldModel(encoder, dynamics, reward_head)
+        td = TensorDict(
+            {
+                "observation": torch.randn(_WM_BATCH, _WM_OBS_DIM),
+                "action": torch.randn(_WM_BATCH, _WM_ACTION_DIM),
+            },
+            batch_size=[_WM_BATCH],
+        )
+        out = wm(td)
+        assert ("next", "encoded", "latent") in out.keys(include_nested=True)
+
+
+class TestWorldModelEnv:
+    """End-to-end tests for :class:`WorldModelEnv`.
+
+    The world model owns prediction; the env owns rollout semantics. Each test
+    builds a ``WorldModelEnv`` around a ``WorldModel`` and drives it through
+    :meth:`EnvBase.rollout` so reset/step/done handling stays consistent with
+    every other TorchRL env.
+    """
+
+    @staticmethod
+    def _make_env(wm: WorldModel) -> WorldModelEnv:
+        return WorldModelEnv(wm, base_env=_SpecOnlyEnv(batch_size=_WM_BATCH))
+
+    def test_rollout_shape(self):
+        wm = _make_linear_world_model()
+        env = self._make_env(wm)
+        start_td = _make_start_td()
+        policy = _constant_policy()
+        horizon = 5
+        out = env.rollout(max_steps=horizon, policy=policy, tensordict=start_td)
+        assert out.shape == torch.Size([_WM_BATCH, horizon])
+
+    def test_rollout_no_done_head(self):
+        """Without a done head, rollouts run for the requested ``max_steps``."""
+        wm = _make_linear_world_model(with_done_head=False)
+        env = self._make_env(wm)
+        start_td = _make_start_td()
+        policy = _constant_policy()
+        out = env.rollout(max_steps=4, policy=policy, tensordict=start_td)
+        assert out.shape == torch.Size([_WM_BATCH, 4])
+
+    def test_rollout_break_when_done(self):
+        """When the done head always predicts done=True, env.rollout stops early."""
+        wm = _make_linear_world_model(with_done_head=True)
+        # Override done head to always emit True so we can verify early stop.
+        wm.done_head = TensorDictModule(
+            lambda x: torch.ones(*x.shape[:-1], 1, dtype=torch.bool),
+            in_keys=[("next", "latent")],
+            out_keys=[("next", "done")],
+        )
+        from tensordict.nn import TensorDictSequential
+
+        wm._step_seq = TensorDictSequential(wm.dynamics, wm.reward_head, wm.done_head)
+
+        env = self._make_env(wm)
+        start_td = _make_start_td()
+        policy = _constant_policy()
+        out = env.rollout(
+            max_steps=10,
+            policy=policy,
+            tensordict=start_td,
+            break_when_any_done=True,
+        )
+        assert out.shape[1] == 1
+
+    def test_rollout_contains_reward(self):
+        wm = _make_linear_world_model()
+        env = self._make_env(wm)
+        start_td = _make_start_td()
+        policy = _constant_policy()
+        out = env.rollout(max_steps=3, policy=policy, tensordict=start_td)
+        assert ("next", "reward") in out.keys(include_nested=True)
+
+    def test_rollout_replay_buffer_compatible(self):
+        """Imagined rollout can be added directly to a TensorDictReplayBuffer."""
+        from torchrl.data import LazyTensorStorage
+
+        wm = _make_linear_world_model()
+        env = self._make_env(wm)
+        start_td = _make_start_td()
+        policy = _constant_policy()
+        rollout_td = env.rollout(max_steps=5, policy=policy, tensordict=start_td)
+        flat = rollout_td.reshape(-1)
+        rb = TensorDictReplayBuffer(storage=LazyTensorStorage(max_size=100))
+        rb.extend(flat)
+        assert len(rb) == flat.batch_size[0]
+
+    def test_reset_requires_latent(self):
+        """WorldModelEnv refuses to reset without an explicit starting latent."""
+        wm = _make_linear_world_model()
+        env = self._make_env(wm)
+        with pytest.raises(RuntimeError, match="initial latent"):
+            env.reset()
